Dra ut SakVelger i egen komponent
